Better error messages for bad application
I think this error message makes sense; because it's a unification
failure I don't want to get into digging into whether the argument or
return type was invalid (both may be invalid simultaneously), but I
think a human can look and understand the error. Of course, this is
only in the small--in a large program I think there's still a ton of
work to do.

New test.