Cleanup type variables after formatting
It turns out that we use the var_temp_other while formatting in order
to track allocated variables, and if we don't clean it up then other
parts of the code see that as a valid pointer and try to walk into it,
which is no good. Remember, always run _CleanupTypeVariables after
doing something that uses var_temp_other.