Stop looking in my closure to get my closure
Turns out we can arrange it so that inside a `let rec` lambda we know
the identifier bound to the closure in r0; this keeps us from
requiring a self-pointer slot in r0.

In the case of factorial, this allows the function to require no
allocations rather than the allocation of a closure with a single
slot.

Also add tracing of all the allocations, and print out the statistics
at the end.